Moves time wheel config into it's own class re #3306
